% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_pknca_options.R
\name{apply_pknca_options}
\alias{apply_pknca_options}
\title{Read settings from a JSON file, and pass to PKNCA.options()}
\usage{
apply_pknca_options(options, nca_parameters = NULL, verbose = TRUE)
}
\arguments{
\item{options}{a list of PKNCA options to be set}

\item{nca_parameters}{list of what parameters to calculate, e.g.
\code{list(auclast=TRUE, cmax=FALSE, ...)}}

\item{verbose}{show verbose output?}
}
\value{
The current PKNCA options.
}
\description{
Read settings from a JSON file, and pass to PKNCA.options()
}
