% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_nca_grouping.R
\name{check_nca_grouping}
\alias{check_nca_grouping}
\title{Check that \code{grouping} specified to PKNCA is likely to be correct
If the grouping results in most groups having too few datapoints to
perform NCA and calculate half-life, then the grouping is likely wrong.}
\usage{
check_nca_grouping(
  data,
  groups,
  dictionary,
  settings = list(),
  threshold = 0.7,
  verbose = TRUE
)
}
\arguments{
\item{data}{dataset in pre-parse format, preferrably created using
\code{create_nca_data()}. Can also be a NONMEM-style dataset with EVID=0|1 column
indicating PK an dosing data.}

\item{groups}{optional grouping variable, e.g. \code{ACTARM}.}

\item{dictionary}{abridged data dictionary for the input dataset, specified
as list object. Requires at least entries for \code{subject_id}, \code{time}, \code{conc},
\code{dose}. Optional entries are: \code{visit}.
If required entries are specified only partially, then default values
(i.e. CDISC) nomenclature will be imputed for missing identifiers.}

\item{settings}{list of settings for NCA. Provided setting names can either
be settings recognized directly by \code{pknca}, or settings that are included in
the map provided by \code{nca_settings_map()} (which are then mapped to \code{pknca}
setting names).}

\item{threshold}{the fraction of data to be allowed to have too few data
points for NCA.}

\item{verbose}{verbose output?}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Check that \code{grouping} specified to PKNCA is likely to be correct
If the grouping results in most groups having too few datapoints to
perform NCA and calculate half-life, then the grouping is likely wrong.
}
