% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_nca_table.R
\name{create_nca_table}
\alias{create_nca_table}
\title{Create a table for NCA results}
\usage{
create_nca_table(
  nca_data,
  parameters = NULL,
  groups = NULL,
  path = NULL,
  format = c("wide", "long"),
  style = c("simple", "pretty"),
  specification = "default",
  description = TRUE,
  prefix = NULL,
  verbose = TRUE
)
}
\arguments{
\item{nca_data}{output data from NCA, generated using \code{run_nca()}}

\item{parameters}{vector of parameters to output, e.g. \code{c("AUCALL", "CMAX", "TMAX")}. If \code{NULL}, defaults to output all available parameters.}

\item{groups}{name of variable in dataset to group statistics by, e.g.
\code{"ACTARM"}}

\item{path}{optional, path to filename to save output table to.}

\item{format}{output as table in wide or long format.}

\item{style}{style of output table. Default is \code{simple}, in which the output
is essentially a regular data.frame that can easily be used for further
parsing. The \code{pretty} style is intended for printing or inclusion in reports:
when multiple rows are in the table for a single parameter or interval, it
will leave blanks on the subsequent rows for parameter and description.}

\item{specification}{specification file with settings for NCA. Needs to match
the \code{specification} that was used in \code{run_nca()} to properly map column
names.}

\item{description}{include description of parameter names into table?}

\item{prefix}{prefix all column names with this, e.g. \code{superpos_}.}

\item{verbose}{Verbose?}
}
\value{
A data.frame or tibble.
}
\description{
Confirms with layout of table commonly used in PK reports, i.e. where
stats are outputted in long format using a single column.
}
