% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nca_individual_fits.R
\name{get_nca_individual_fits}
\alias{get_nca_individual_fits}
\title{Extract individul fit data from a PKNCA results object, and return as
data.frame convenient for plotting.}
\usage{
get_nca_individual_fits(nca_obj = NULL, n_grid_points = 6, dictionary = NULL)
}
\arguments{
\item{nca_obj}{object returned from PKNCA with NCA results.}

\item{n_grid_points}{number of grid points to simulate for the fitted line.
Usually 40 is more than enough.}

\item{dictionary}{list object, mapping column names to expected variables.
Any specified elements, if present in \code{nca_obj$data$conc$data}, will be
exported in the output object.}

\item{exclusions}{a vector of points to remove ## TODO}
}
\description{
Extract individul fit data from a PKNCA results object, and return as
data.frame convenient for plotting.
}
