% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_nca.R
\name{get_nca_units}
\alias{get_nca_units}
\title{Get a table of units from a raw PKNCA object}
\usage{
get_nca_units(data, units_col = "PPSTRESU")
}
\arguments{
\item{data}{PKNCA output object}

\item{units_col}{the name of the column holding the units, default is
\code{PPSTRESU}}
}
\value{
data.frame with \code{name} and \code{unit} columns
}
\description{
Get a table of units from a raw PKNCA object
}
