% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nominal_timepoints.R
\name{get_nominal_timepoints}
\alias{get_nominal_timepoints}
\title{Get nominal timepoints based on vector of timepoints}
\usage{
get_nominal_timepoints(t, adjust = 0.5, ...)
}
\arguments{
\item{t}{vector of timepoints
Get nominal timepoints for a given vector of timepoints}

\item{...}{arguments passed to \code{stats::density} function, e.g \code{bw} or
\code{adjust}. The easiest way to adjust the sensitivity to peaks in the data
is to use \code{adjust}. A factor of 0.5 often works well for PK data.}
}
\value{
a vector of approximate nominal timepoints estimated from the data
}
\description{
Get nominal timepoints based on vector of timepoints
}
