% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_nca_table.R
\name{get_parameter_names_from_nca_data}
\alias{get_parameter_names_from_nca_data}
\title{Get parameter names from a run_nca() output object}
\usage{
get_parameter_names_from_nca_data(nca_data)
}
\arguments{
\item{nca_data}{output data from NCA, generated using \code{run_nca()}}
}
\description{
Get parameter names from a run_nca() output object
}
