% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_nca.R
\name{get_pk_parameters_from_pknca_options}
\alias{get_pk_parameters_from_pknca_options}
\title{When PKNCA::PKNCA.options() is invoked without arguments, it will
return a list of its configuration. Among the options is the data.frame
\code{single.dose.aucs} which will list the exported PK parameters for both
the AUC0-24 and AUCinf computations. Wherever the parameters are set to TRUE
PKNCA will compute those and include in output. So in that case we can use
those parameters as well and should make them available in the output from
\code{run_nca()}.}
\usage{
get_pk_parameters_from_pknca_options(options = NULL, no_dots = FALSE)
}
\arguments{
\item{options}{parameters for interval specification for NCA}

\item{no_dots}{if \code{TRUE} (default) will replace any dots in parameter names
(e.g. \code{aucinf.obs}) with underscores (\code{aucinf_obs}).}
}
\description{
When PKNCA::PKNCA.options() is invoked without arguments, it will
return a list of its configuration. Among the options is the data.frame
\code{single.dose.aucs} which will list the exported PK parameters for both
the AUC0-24 and AUCinf computations. Wherever the parameters are set to TRUE
PKNCA will compute those and include in output. So in that case we can use
those parameters as well and should make them available in the output from
\code{run_nca()}.
}
