% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_treatment_sequence.R
\name{get_treatment_sequence}
\alias{get_treatment_sequence}
\title{Function to generate a sequence variable for e.g. a crossover trial design}
\usage{
get_treatment_sequence(data, dictionary, variable = "treatment", groups = NULL)
}
\arguments{
\item{data}{dataset in pre-parse format, preferrably created using
\code{create_nca_data()}. Can also be a NONMEM-style dataset with EVID=0|1 column
indicating PK an dosing data.}

\item{dictionary}{abridged data dictionary for the input dataset, specified
as list object. Requires at least entries for \code{subject_id}, \code{time}, \code{conc},
\code{dose}. Optional entries are: \code{visit}.
If required entries are specified only partially, then default values
(i.e. CDISC) nomenclature will be imputed for missing identifiers.}

\item{variable}{variable name to base the sequence detection on}

\item{groups}{optional grouping variable, e.g. \code{ACTARM}.}
}
\value{
a data frame with the subject_id, sequence number, and full sequence
description.
}
\description{
Function to generate a sequence variable for e.g. a crossover trial design
}
