% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_pknca_options.R
\name{is_set_to_missing}
\alias{is_set_to_missing}
\title{Is value meant to be \code{NA}?}
\usage{
is_set_to_missing(x)
}
\arguments{
\item{x}{value}
}
\description{
Is value meant to be \code{NA}?
}
