% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_nca_settings.R
\name{map_nca_settings}
\alias{map_nca_settings}
\title{Map user-provided NCA settings to NCA engine (e.g. PKNCA)}
\usage{
map_nca_settings(settings = NULL, mapping = "pknca")
}
\arguments{
\item{settings}{list of settings, potentially to be remapped}

\item{mapping}{name of mapping to use, default is \code{pknca}. Mappings are
stored within package in folder \code{md/nca}.}
}
\description{
Map user-provided NCA settings to NCA engine (e.g. PKNCA)
}
