% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_add_auctau.R
\name{nca_add_auctau}
\alias{nca_add_auctau}
\title{Add AUCtau to the NCA output, if multi-dose PK data}
\usage{
nca_add_auctau(output, dictionary, verbose = FALSE)
}
\arguments{
\item{output}{PKNCA output object}

\item{dictionary}{data dictionary}

\item{verbose}{verbose output?}
}
\value{
NCA output data with \code{auctau} added.
}
\description{
Will copy \code{auclast} to \code{auctau} in the NCA output, for
those intervals where the end of the interval is <Inf.
}
\details{
It will first check if the NCA was indeed multi-dose,
as this is not relevant variable for single-dose data.
}
