% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nca_admiral}
\alias{nca_admiral}
\title{Example data for NCA}
\format{
A data frame with 2352 rows and
48 columns.
}
\usage{
nca_admiral
}
\description{
Example data from admiral dataset for use in NCA
}
\keyword{datasets}
