% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_post.R
\name{nca_post_accumulation}
\alias{nca_post_accumulation}
\title{NCA post-processing function to calculate accumulation rate}
\usage{
nca_post_accumulation(
  data,
  dictionary,
  options = list(parameters = c("auctau", "cmax")),
  groups = NULL,
  verbose = FALSE
)
}
\arguments{
\item{data}{NCA output data}

\item{dictionary}{data dictionary}

\item{options}{options for this function:
\itemize{
\item \code{parameters}: vector of parameter names to calculate accumulation ratios for.
}}

\item{groups}{grouping variable}

\item{verbose}{verbosity}
}
\description{
Will only be able to calculate accumulation ratio when data is multi-
interval.
}
