% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_post.R
\name{nca_post_bioavailability}
\alias{nca_post_bioavailability}
\title{NCA post-processing function to calculate relative bioavailability}
\usage{
nca_post_bioavailability(
  data,
  dictionary,
  options = list(parameters = c("auclast", "cmax"), arm = NULL, reference = NULL),
  groups = NULL,
  verbose = FALSE
)
}
\arguments{
\item{data}{NCA output data}

\item{dictionary}{data dictionary}

\item{options}{options for this function:
\itemize{
\item \code{parameters}: vector of parameter names to calculate ratio for.
\item \code{arm}: the column name used to indicate the varios arms of the study, e.g.
fasting / fed, or iv / oral, etc. Can have multiple test arms, but only one
\code{reference}.
\item \code{reference}: the label of the reference arm in \code{group} column, e.g. "fasted"
}}

\item{groups}{grouping variable}

\item{verbose}{verbosity}
}
\description{
Will only be able to calculate relative bioavailability when data
has a grouper to distinguish the test and reference arms, e.g.
}
