% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_exclusions.R
\name{parse_exclusions}
\alias{parse_exclusions}
\title{Helper function to get a mask based on a list of vectors with identifiers.
This function is used to gather exclusions in the \code{run_nca()} function.}
\usage{
parse_exclusions(data, exclusions)
}
\arguments{
\item{data}{data.frame}

\item{exclusions}{a named list of lists, with the names corresponding
to columns in the dataset. Each sub-list should contain vectors \code{id} and
\code{reason}. E.g.

\code{list( SAMPLEID = list( id = c("017011111-20120908T120000", "017011028-20130719T010000", "017011033-20140318T040000"),  reason = c("unexplained anomaly", "manual lambda-z check", "assay error") ) )}}
}
\value{
a list with: the mask (vector of T/F) that indicates the rows in the data to
exclude, and a vector of reasons for exclusions (NA if not excluded)
}
\description{
Helper function to get a mask based on a list of vectors with identifiers.
This function is used to gather exclusions in the \code{run_nca()} function.
}
