% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_pknca_options.R
\name{parse_pknca_opts_from_spec}
\alias{parse_pknca_opts_from_spec}
\title{Get the PKNCA options from a spec file, and parse into
expected format for \code{PKNCA::PKNCA.options()}}
\usage{
parse_pknca_opts_from_spec(opts)
}
\arguments{
\item{opts}{list of options for PKNCA}
}
\value{
A formatted list of arguments to be passed to \code{PKNCA::PKNCA.options()}.
}
\description{
Get the PKNCA options from a spec file, and parse into
expected format for \code{PKNCA::PKNCA.options()}
}
