% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKNCA.extra-package.R
\docType{package}
\name{PKNCA.extra-package}
\alias{PKNCA.extra}
\alias{PKNCA.extra-package}
\title{PKNCA.extra: Functionality for non-compartmental analysis using PKNCA}
\description{
Provides functions to perform NCAs using the PKNCA package as the underlying computational engine. The PKNCA package provides a great and stable computational engine, but is somewhat limited (by design) in the features it provides for data (pre-)processing and interpretation of the NCA results. This package offers some of that functionality.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://insightrx.github.io/PKNCA.extra/}
}

}
\author{
\strong{Maintainer}: Ron Keizer \email{ron@insight-rx.com}

Other contributors:
\itemize{
  \item Michael McCarthy \email{michael.mccarthy@insight-rx.com} [contributor]
  \item InsightRX [copyright holder]
}

}
\keyword{internal}
