% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_spec_from_file.R
\name{read_spec_from_file}
\alias{read_spec_from_file}
\title{Read specification from metadata file}
\usage{
read_spec_from_file(spec_file)
}
\arguments{
\item{spec_file}{JSON file with specifications for NCA. See example files
for more info.}
}
\description{
Read specification from metadata file
}
