% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_nca_settings.R
\name{remap_core}
\alias{remap_core}
\title{Core remapping function}
\usage{
remap_core(settings, m, mapped_name, mapped_option)
}
\arguments{
\item{settings}{TODO}

\item{m}{TODO}

\item{mapped_name}{TODO}

\item{mapped_option}{TODO}
}
\description{
Core remapping function
}
