% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remap_nca_output.R
\name{remap_nca_output}
\alias{remap_nca_output}
\title{Function to export output from PKNCA to CDISC- or other mappings}
\usage{
remap_nca_output(data, mapping = NULL, exclude_unmapped = TRUE, keep = c())
}
\arguments{
\item{data}{output object from \code{PKNCA::pk.nca()}}

\item{mapping}{a list mapping PKNCA parameters to custom parameter names,
e.g. \verb{list(CMAX_custom = "CMAX", "auc_last = "AUCLAST")}. PKNCA parameter
names are list values, new parameter names are list names.}

\item{exclude_unmapped}{exclude unmapped variables?}

\item{keep}{parameters to keep when remapping and excluding unmapped
parameters.}
}
\value{
a data.frame with column names matching the desired mapping
}
\description{
This function could be temporary in this package, PKNCA may support
remapping in the future, at least to CDISC format.
}
\examples{
\dontrun{
nca <- run_nca(...)
nca_cdisc <- remap_nca_output(nca, "cdisc")
}

}
