% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_nca_superposition.R
\name{run_nca_superposition}
\alias{run_nca_superposition}
\title{Run an NCA on superpositioned data, either just extrapolated to steady state
for the current dose levels, or using linear extrapolation to a different
dose and interval.}
\usage{
run_nca_superposition(
  data,
  dose_data = NULL,
  regimen,
  dictionary,
  parameters = c("auclast", "cmax"),
  groups = c(),
  verbose = TRUE
)
}
\arguments{
\item{data}{dataset in pre-parse format, preferrably created using
\code{create_nca_data()}. Can also be a NONMEM-style dataset with EVID=0|1 column
indicating PK an dosing data.}

\item{dose_data}{optional, a data.frame with all dosing info. Useful for
multi-dose NCAs. Alternative to specifying a NONMEM-style datasett to \code{data}.}

\item{regimen}{a list that specifies at least the \code{interval}, and
potentially \code{dose}. E.g. \code{list(interval = 8)} to calculate steady state
concentration data using superposition. If a \code{dose} is also specified,
it will scale the superposition data linearly compared to the original
input \code{dose}, e.g. \code{list(dose = 500, interval = 12)}.}

\item{dictionary}{abridged data dictionary for the input dataset, specified
as list object. Requires at least entries for \code{subject_id}, \code{time}, \code{conc},
\code{dose}. Optional entries are: \code{visit}.
If required entries are specified only partially, then default values
(i.e. CDISC) nomenclature will be imputed for missing identifiers.}

\item{groups}{optional grouping variable, e.g. \code{ACTARM}.}

\item{verbose}{verbose output?}
}
\value{
a PKNCA output object based on superposition data
}
\description{
Run an NCA on superpositioned data, either just extrapolated to steady state
for the current dose levels, or using linear extrapolation to a different
dose and interval.
}
