% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_nca.R
\name{set_pknca_interval_parameters}
\alias{set_pknca_interval_parameters}
\title{Creates a data.frame specifying start and end time, and stores it as option
\code{single.dose.auc} in PKNCA.options()}
\usage{
set_pknca_interval_parameters(time, parameters)
}
\arguments{
\item{time}{description...}

\item{parameters}{list of parameters to calculate, indicated by TRUE and
FALSE, e.g. \code{list(auc.inf.pred=TRUE, cmax.pred=FALSE, ...)}}
}
\description{
Creates a data.frame specifying start and end time, and stores it as option
\code{single.dose.auc} in PKNCA.options()
}
