% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_nca.R
\name{run_nca}
\alias{run_nca}
\title{Run a non-compartmental analysis}
\usage{
run_nca(
  data,
  dose_data = NULL,
  specification = "default",
  dictionary = NULL,
  settings = NULL,
  blq = "<",
  groups = NULL,
  include_cols = NULL,
  intervals = data.frame(start = 0, end = Inf),
  partial_auc = NULL,
  exclude_points = NULL,
  exclude_subjects = NULL,
  sequence_from = NULL,
  add_auctau = TRUE,
  post = list(accumulation = list(parameters = c("auctau", "cmax"))),
  no_dots = TRUE,
  path = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{data}{dataset in pre-parse format, preferrably created using
\code{create_nca_data()}. Can also be a NONMEM-style dataset with EVID=0|1 column
indicating PK an dosing data.}

\item{dose_data}{optional, a data.frame with all dosing info. Useful for
multi-dose NCAs. Alternative to specifying a NONMEM-style datasett to \code{data}.}

\item{specification}{use specific options for the NCA and for mapping of
input / output data.}

\item{dictionary}{abridged data dictionary for the input dataset, specified
as list object. Requires at least entries for \code{subject_id}, \code{time}, \code{conc},
\code{dose}. Optional entries are: \code{visit}.
If required entries are specified only partially, then default values
(i.e. CDISC) nomenclature will be imputed for missing identifiers.}

\item{settings}{list of settings for NCA. Provided setting names can either
be settings recognized directly by \code{pknca}, or settings that are included in
the map provided by \code{nca_settings_map()} (which are then mapped to \code{pknca}
setting names).}

\item{blq}{optional, string to match in concentration data, e.g. "<" or
"BLQ>. If \code{conc} column in the data is \code{character} class, then if \code{blq} is
not NULL, it will detect any values that match this string and set those
values to zero. If NULL, or \code{conc} data is numeric, the data will be used
as-is.}

\item{groups}{optional grouping variable, e.g. \code{ACTARM}.}

\item{include_cols}{Optional. Character vector of column names in \code{data} that
should be merged in with NCA results (including only a single value for each
row). Merge will happen by subject ID, any supplied groupings, and
\code{nca_start} time.}

\item{intervals}{specify time intervals for NCA calculations. The passed argument
should be a data.frame, with at least the \code{start} and \code{end} time for each
row that specifies the interval. \code{Inf} is allowed to specify the end of the
interval. Example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{intervals = bind_rows(
   c(start = 0, end = 24), 
   c(start = 120, end = Inf)
)
}\if{html}{\out{</div>}}

The \code{run_nca()} function will compute all parameters that are specified in
\code{settings()} or in the default settings.}

\item{partial_auc}{a data.frame containing at least the \code{start} and \code{end}
time for each row that specifies the start- and endtime for the partial AUC
calculation. At least the AUC_0-t (\code{auclast}) will be calculated for each
row. Any other parameters specified in the data.frame as column will also be
calculated, e.g. \code{cmax=TRUE}. Currently, only a single row can be supplied
as argument.}

\item{exclude_points}{list of points to exclude, based on name(s) in the
list that should match column name in the input dataset.
E.g. \code{list("sample_id" = c("12345", "23456"))}. This could similarly be used
to exclude entire subjects, e.g. \verb{list("subject_id = c("103, "105")}}

\item{exclude_subjects}{description...}

\item{sequence_from}{optional. Argument should specify the column name
that indicates the treatment e.g. for food-effect cross-over trials. It
will then add sequence group and sequence description columns to the
output data.}

\item{add_auctau}{compute AUCtau whenever relevant? TRUE (default) / FALSE.
For multi-dose PK data, this will add the parameter \code{auc_tau} to the NCA
output data as the AUC for each interval. In essence, it will just make a
copy of the AUClast for the interval. See exact logic in function
help page.}

\item{post}{named list of lists, indicating which hard-coded
post-processing steps to include, and options for each step.
Possible post-processing steps include:
\itemize{
\item \code{accumulation}: if NCA calculated for multiple non-overlapping intervals,
will add accumulation ratios for the parameters listed in the \code{parameters}
argument to the NCA output table for each interval.
}

Other options may be added in future. Default is:
\code{list("accumulation" = list("parameters" = c("auclast", "cmax")))}.}

\item{no_dots}{if \code{TRUE} (default) will replace any dots in parameter names
(e.g. \code{aucinf.obs}) with underscores (\code{aucinf_obs}).}

\item{path}{path to file to store output object from PKNCA as RDS (optional)}

\item{verbose}{verbose output?}

\item{...}{parameters passed to `PKNCA::pk.nca()`` function.}
}
\description{
This function runs an NCA on a pre-parsed dataset. The NCA will be performed
using the PKNCA package. The NCA output object will be saved to an
rds file if \code{path} is specified.
}
\details{
There are four possible types of input datasets that can be used to run the
NCA:
\enumerate{
\item Using a separate data.frame for PK data (\code{data}) and dose data
(\code{dose_data}), with each PK or dosing event a row in the respective datasedt.
Can be used for both single and multi-dose data.
\item Using a single data.frame for PK and dose data (\code{data}), with event type
separated using an evid column (specified in \code{dictionary}). PK data rows
should be set to \code{evid=0}, and doses to \code{evid=1}. This is essentially a
NONMEM-style dataset. Rows with other \code{evid} values will be ignored.
Can be used for both single and multi-dose data, the dosing occasions will
be determined from the dosing events. If an \code{evid} column is present in
\code{data}, it will not attempt to use approach 3.
\item Using a single data.frame (\code{data}) with PK data as rows, and a column for
dose (with colum name specified in \code{dictionary}). A \code{dose_data} data.frame
will be constructed from the dose column. Can only be used for single-dose
data, unless \code{groups} are used to separate the dosing occasions.
\item Using a single data.frame (\code{data}) with PK data as rows, but no dosing
info in the data. NCA can still compute AUC and Cmax etc, but parameters that
require the dose such as CL, V, etc will be returned as NA. Can only be used
for single dose, unless \code{groups} are used to separate the dosing occasions.
}
}
\examples{
\dontrun{
run_nca(
  data = data,
  path = "./outputs/nca.rds"
)
}
}
